+++ crates/core-common/src/tests/common.rs
@@ -2,6 +2,47 @@
 
 use stash::stash::Stash;
 
+/// Macro wrapping u64 into Option<LocalId> for easier model definition.
+#[macro_export]
+macro_rules! lid {
+    ($id:expr) => {{
+        use crate::datatypes::LocalId;
+        Some(LocalId::from($id))
+    }};
+}
+
+/// Macro wrapping &str into Option<RemoteId> for easier model definition.
+/// Since it calls .into() on the RemoteId, it allows creation of Option<LabelId> as well.
+#[macro_export]
+macro_rules! rid {
+    ($id:expr) => {{
+        use crate::datatypes::RemoteId;
+        Some(RemoteId::from($id).into())
+    }};
+}
+
+#[macro_export]
+macro_rules! contact {
+    ($($field:tt)*) => {{
+        use crate::models::Contact;
+        Contact {
+            $($field)*,
+            ..Default::default()
+        }
+    }};
+}
+
+#[macro_export]
+macro_rules! contact_email {
+    ($($field:tt)*) => {{
+        use crate::models::ContactEmail;
+        ContactEmail {
+            $($field)*,
+            ..Default::default()
+        }
+    }};
+}
+
 pub async fn new_core_test_connection() -> Stash {
     use std::io::stdout;
     use tracing::subscriber::set_global_default;