+++ crates/core-common/src/tests/utils.rs
@@ -1,0 +1,21 @@
+use super::*;
+use test_case::test_case;
+
+#[test_case("a" => "A")]
+#[test_case("B" => "B")]
+#[test_case("1" => "1")]
+#[test_case("y̆es" => "Y̆")]
+#[test_case("🗻∈🌏" => "🗻")]
+#[test_case("\"This is a quote\"" => "\"")]
+#[test_case("🧑‍🔬 Doctor Rebecca" => "🧑‍🔬")]
+fn test_first_grapheme_uppercase(s: &str) -> String {
+    first_grapheme_upppercase(s).unwrap_or_default()
+}
+
+#[test]
+fn test_proton_color() {
+    assert_eq!(proton_color("John Doe"), "#3C8B8C");
+    assert_eq!(proton_color("Jane Doe"), "#0F735A");
+    assert_eq!(proton_color("Test"), "#A839A4");
+    assert_eq!(proton_color(""), "#0F735A");
+}